+++ index.php
@@ -6,5 +6,7 @@ require_once('vendor/autoload.php');
 
 use Asylamba\Classes\Worker\Application;
 
+define("P_TYPE", 'app');
+
 $application = new Application();
 $application->boot();