+++ system/Classes/Task/CyclicTask.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Asylamba\Classes\Task;
+
+class CyclicTask extends Task
+{
+	/**
+	 * {@inheritdoc}
+	 */
+    public function getType()
+	{
+		return self::TYPE_CYCLIC;
+	}
+}