+++ system/Classes/Task/RealTimeTask.php
@@ -1,0 +1,64 @@
+<?php
+
+namespace Asylamba\Classes\Task;
+
+class RealTimeTask extends Task
+{
+    /** @var int **/
+    protected $objectId;
+    /** @var string **/
+    protected $date;
+	
+	/**
+	 * {@inheritdoc}
+	 */
+    public function getType()
+	{
+		return self::TYPE_REALTIME;
+	}
+    
+    /**
+     * @param int $id
+     * @return $this
+     */
+    public function setObjectId($id)
+    {
+        $this->objectId = $id;
+        
+        return $this;
+    }
+    
+    /**
+     * @return int
+     */
+    public function getObjectId()
+    {
+        return $this->objectId;
+    }
+    
+    /**
+     * @param string $date
+     * @return $this
+     */
+    public function setDate($date)
+    {
+        $this->date = $date;
+        
+        return $this;
+    }
+    
+    /**
+     * @return string
+     */
+    public function getDate()
+    {
+        return $this->date;
+    }
+	
+	public function jsonSerialize() {
+		return array_merge(parent::jsonSerialize(), [
+			'object_id' => $this->objectId,
+			'date' => $this->date
+		]);
+	}
+}