+++ system/Classes/Task/TaskManager.php
@@ -1,0 +1,143 @@
+<?php
+
+namespace Asylamba\Classes\Task;
+
+use Asylamba\Classes\DependencyInjection\Container;
+
+use Asylamba\Classes\Process\Process;
+
+class TaskManager
+{
+	/** @var Container **/
+	protected $container;
+    
+	/**
+	 * @param Container $container
+	 */
+    public function __construct(Container $container)
+    {
+		$this->container = $container;
+    }
+	
+	/**
+	 * @param array $data
+	 * @return Task
+	 */
+	public function createTaskFromData($data)
+	{
+		switch ($data['type']) {
+			case Task::TYPE_TECHNICAL:
+				return $this->createTechnicalTask($data['manager'], $data['method'], $data['id']);
+			case Task::TYPE_REALTIME:
+				return $this->createRealTimeTask($data['manager'], $data['method'], $data['object_id'], $data['date'], $data['id']);
+			case Task::TYPE_CYCLIC:
+				return $this->createCyclicTask($data['manager'], $data['method'], $data['id']);
+		}
+	}
+    
+    /**
+     * @param string $manager
+     * @param string $method
+     * @param int $objectId
+     * @param string $date
+     * @return Task
+     */
+    public function createRealTimeTask($manager, $method, $objectId, $date, $id = null)
+    {
+        return
+            (new RealTimeTask())
+            ->setId((($id !== null) ? $id : $this->generateId()))
+            ->setManager($manager)
+            ->setMethod($method)
+            ->setObjectId($objectId)
+            ->setDate($date)
+        ;
+    }
+    
+    /**
+     * @param string $manager
+     * @param string $method
+     * @return Task
+     */
+    public function createCyclicTask($manager, $method, $id = null)
+    {
+        return
+            (new CyclicTask())
+            ->setId((($id !== null) ? $id : $this->generateId()))
+            ->setManager($manager)
+            ->setMethod($method)
+        ;
+    }
+    
+    /**
+     * @param string $manager
+     * @param string $method
+     * @return Task
+     */
+    public function createTechnicalTask($manager, $method, $id = null)
+    {
+        return
+            (new TechnicalTask())
+            ->setId((($id !== null) ? $id : $this->generateId()))
+            ->setManager($manager)
+            ->setMethod($method)
+        ;
+    }
+    
+	/**
+	 * @param \Asylamba\Classes\Task\Task $task
+	 * @return array
+	 */
+    public function perform(Task $task)
+    {
+        try {
+            // Get the manager from the container and then execute the given method with its arguments
+            call_user_func_array(
+                [$this->container->get($task->getManager()), $task->getMethod()],
+                ($task instanceof RealTimeTask) ? [$task->getObjectId()] : []
+            );
+            return ['success' => true, 'task' => $task];
+        } catch (\Exception $ex) {
+            return [
+                'success' => false,
+                'error' => [
+                    'message' => $ex->getMessage(),
+                    'file' => $ex->getFile(),
+                    'line' => $ex->getLine()
+                ]
+            ];
+        } catch (\Error $err) {
+            return [
+                'success' => false,
+                'error' => [
+                    'message' => $err->getMessage(),
+                    'file' => $err->getFile(),
+                    'line' => $err->getLine()
+                ]
+            ];
+        }
+    }
+	
+	/**
+	 * @param Process $process
+	 * @param array $data
+	 */
+	public function validateTask(Process $process, $data)
+	{
+		if ($data['success'] === false) {
+			return;
+		}
+		$task = $this->createTaskFromData($data['task']);
+		
+		$task->setTime((float) $data['time']);
+		
+		$process->removeTask($task);
+		
+		$this->container->get('load_balancer')->storeStats($task);
+	}
+    
+    public function generateId()
+    {
+        return uniqid('process_');
+    }
+}