+++ system/Modules/Ares/Manager/CommanderManager.php
@@ -21,7 +21,7 @@ use Asylamba\Modules\Zeus\Manager\PlayerManager;
 use Asylamba\Modules\Zeus\Manager\PlayerBonusManager;
 use Asylamba\Modules\Gaia\Manager\PlaceManager;
 use Asylamba\Modules\Demeter\Manager\ColorManager;
-use Asylamba\Modules\Athena\Manager\RecyclingMissionManager;
+use Asylamba\Modules\Hermes\Manager\NotificationManager;
 use Asylamba\Classes\Library\Session\SessionWrapper;
 use Asylamba\Classes\Container\ArrayList;
 
@@ -58,8 +58,8 @@ class CommanderManager
 	protected $placeManager;
 	/** @var ColorManager **/
 	protected $colorManager;
-	/** @var RecyclingMissionManager **/
-	protected $recyclingMissionManager;
+	/** @var NotificationManager **/
+	protected $notificationManager;
 	/** @var Session **/
 	protected $session;
 	/** @var RealTimeActionScheduler **/
@@ -85,7 +85,7 @@ class CommanderManager
 	 * @param PlayerBonusManager $playerBonusManager
 	 * @param PlaceManager $placeManager
 	 * @param ColorManager $colorManager
-	 * @param RecyclingMissionManager $recyclingMissionManager
+	 * @param NotificationManager $notificationManager
 	 * @param SessionWrapper $session
 	 * @param RealTimeActionScheduler $scheduler
 	 * @param EventDispatcher $eventDispatcher
@@ -100,7 +100,7 @@ class CommanderManager
 		PlayerBonusManager $playerBonusManager,
 		PlaceManager $placeManager,
 		ColorManager $colorManager,
-		RecyclingMissionManager $recyclingMissionManager,
+		NotificationManager $notificationManager,
 		SessionWrapper $session,
 		RealTimeActionScheduler $scheduler,
 		EventDispatcher $eventDispatcher,
@@ -114,7 +114,7 @@ class CommanderManager
 		$this->playerBonusManager = $playerBonusManager;
 		$this->placeManager = $placeManager;
 		$this->colorManager = $colorManager;
-		$this->recyclingMissionManager = $recyclingMissionManager;
+		$this->notificationManager = $notificationManager;
 		$this->session = $session;
 		$this->scheduler = $scheduler;
 		$this->eventDispatcher = $eventDispatcher;
@@ -183,13 +183,7 @@ class CommanderManager
 	 */
 	public function getIncomingAttacks($places)
 	{
-		$commanders = $this->entityManager->getRepository(Commander::class)->getIncomingAttacks($places);
-		
-		foreach($commanders as $commander) {
-			$this->uCommander($commander);
-		}
-		
-		return $commanders;
+		return $this->entityManager->getRepository(Commander::class)->getIncomingAttacks($places);
 	}
 	
 	/**
@@ -198,13 +192,7 @@ class CommanderManager
 	 */
 	public function getIncomingCommanders($place)
 	{
-		$commanders = $this->entityManager->getRepository(Commander::class)->getIncomingCommanders($place);
-		
-		foreach($commanders as $commander) {
-			$this->uCommander($commander);
-		}
-		
-		return $commanders;
+		return $this->entityManager->getRepository(Commander::class)->getIncomingCommanders($place);
 	}
 	
 	public function scheduleMovements()
@@ -526,8 +514,7 @@ class CommanderManager
 
 			# changer rBase commander
 			$commander->rBase = $place->id;
-			$commander->travelType = NULL;
-			$commander->statement = Commander::AFFECTED;
+			$this->endTravel($commander, Commander::AFFECTED);
 
 			# ajouter à la place le commandant
 			$place->commanders[] = $commander;
@@ -537,8 +524,7 @@ class CommanderManager
 		} else {
 			# changer rBase commander
 			$commander->rBase = $place->id;
-			$commander->travelType = NULL;
-			$commander->statement = Commander::RESERVE;
+			$this->endTravel($commander, Commander::RESERVE);
 
 			$this->emptySquadrons($commander);
 
@@ -579,9 +565,6 @@ class CommanderManager
 		# si la planète est vide
 		if ($place->rPlayer == NULL) {
 			LiveReport::$isLegal = Report::LEGAL;
-
-			$commander->travelType = NULL;
-			$commander->travelLength = NULL;
 
 			# planète vide : faire un combat
 			$this->startFight($place, $commander, $commanderPlayer);
@@ -628,9 +611,6 @@ class CommanderManager
 			# planète à joueur : si $this->rColor != commandant->rColor
 			# si il peut l'attaquer
 			if (($place->playerColor != $commander->getPlayerColor() && $place->playerLevel > 1 && $commanderColor->colorLink[$place->playerColor] != Color::ALLY) || ($place->playerColor == 0)) {
-				$commander->travelType = NULL;
-				$commander->travelLength = NULL;
-
 				$dCommanders = array();
 				foreach ($place->commanders AS $dCommander) {
 					if ($dCommander->statement == Commander::AFFECTED && $dCommander->line == 1) {
@@ -717,9 +697,6 @@ class CommanderManager
 		$this->playerBonusManager->load($playerBonus);
 		# conquete
 		if ($place->rPlayer != NULL) {
-			$commander->travelType = NULL;
-			$commander->travelLength = NULL;
-
 			if (($place->playerColor != $commander->getPlayerColor() && $place->playerLevel > 3 && $commanderColor->colorLink[$place->playerColor] != Color::ALLY) || ($place->playerColor == 0)) {
 				$tempCom = array();
 
@@ -806,17 +783,12 @@ class CommanderManager
 					$place->playerColor = $commander->playerColor;
 					$place->rPlayer = $commander->rPlayer;
 
-					$S_REM_C1 = $this->recyclingMissionManager->getCurrentSession();
-					$this->recyclingMissionManager->newSession();
-					$this->recyclingMissionManager->load(array('rBase' => $place->id));
-					$S_REM_C2 = $this->recyclingMissionManager->getCurrentSession();
-					$this->recyclingMissionManager->changeSession($S_REM_C1);
 					# changer l'appartenance de la base (et de la place)
-					$this->orbitalBaseManager->changeOwnerById($place->id, $placeBase, $commander->getRPlayer(), $recyclingSession, $baseCommanders);
+					$this->orbitalBaseManager->changeOwnerById($place->id, $placeBase, $commander->getRPlayer(), $baseCommanders);
 					$place->commanders[] = $commander;
 
 					$commander->rBase = $place->id;
-					$commander->statement = Commander::AFFECTED;
+					$this->endTravel($commander, Commander::AFFECTED);
 					$commander->line = 2;
 					
 					$this->eventDispatcher->dispatch(new PlaceOwnerChangeEvent($place));
@@ -851,9 +823,6 @@ class CommanderManager
 
 		# colonisation
 		} else {
-			$commander->travelType = NULL;
-			$commander->travelLength = NULL;
-
 			# faire un combat
 			LiveReport::$type = Commander::COLO;
 			LiveReport::$dFight = $commander->dArrival;
@@ -885,7 +854,7 @@ class CommanderManager
 
 				# attibuer le commander à la place
 				$commander->rBase = $place->id;
-				$commander->statement = Commander::AFFECTED;
+				$this->endTravel($commander, Commander::AFFECTED);
 				$commander->line = 2;
 
 				# ajout de la place en session
@@ -937,13 +906,8 @@ class CommanderManager
 		$commander = $this->get($commanderId);
 		$place = $this->placeManager->get($commander->rDestinationPlace);
 		$commanderBase = $this->orbitalBaseManager->get($commander->rBase);
-		$commander->travelType = null;
-		$commander->travelLength = null;
-		$commander->dStart = null;
-		$commander->dArrival = null;
-		$commander->rStartPlace = null;
-		$commander->rDestinationPlace = null;
-		$commander->statement = Commander::AFFECTED;
+		
+		$this->endTravel($commander, Commander::AFFECTED);
 
 		$this->placeManager->sendNotif($place, Place::COMEBACK, $commander);
 
@@ -952,6 +916,21 @@ class CommanderManager
 			$commander->resources = 0;
 		}
 		$this->entityManager->flush($commander);
+	}
+	
+	/**
+	 * @param Commander $commander
+	 * @param string $statement
+	 */
+	public function endTravel(Commander $commander, $statement)
+	{
+		$commander->travelType = null;
+		$commander->travelLength = null;
+		$commander->dStart = null;
+		$commander->dArrival = null;
+		$commander->rStartPlace = null;
+		$commander->rDestinationPlace = null;
+		$commander->statement = $statement;
 	}
 
 	# HELPER